backend/requirements.txt
Django==5.1.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djoser==2.3.1
drf-nested-routers==0.94.2
gunicorn==23.0.0
psycopg2==2.9.10
python-decouple==3.8
whitenoise==6.9.0
